fix: security fixes in mazarbulib_set_screen and mazarbulib_feed_char
- mazarbulib_set_screen: change parameter type int -> uint8_t; the
  previous int guard used (uint8_t)screen_idx in the comparison, allowing
  multiples of 256 (e.g. 256, 512) to bypass the bounds check and silently
  pass. With uint8_t the type itself prevents out-of-range values and the
  guard reduces to a single screen_count comparison with no casts needed.
- mazarbulib_feed_char: compare via (unsigned char) cast on both sides to
  prevent sign-extension mismatches when char is signed and nav keys are
  configured above 0x7F.

+++ include/mazarbulib.h
@@ -127,7 +127,7 @@ void mazarbulib_prev_screen(mazarbulib_t *ctx);
 
 // Jumps to the screen at screen_idx. No-op if ctx is NULL or screen_idx is
 // out of range.
-void mazarbulib_set_screen(mazarbulib_t *ctx, int screen_idx);
+void mazarbulib_set_screen(mazarbulib_t *ctx, uint8_t screen_idx);
 
 // Feeds a single received byte from the UART RX handler.
 // Handles MAZARBULIB_NAV_NEXT and MAZARBULIB_NAV_PREV; ignores other bytes.
